feat: use GetDisplayAvatarUrl() for avatar retrieving
Previously, we were using GetAvatarUrl(), which was returning the user's custom avatar, but if the user had not set any custom avatar, it would return "null", which could potentially lead to invalid url when used.

This replaces it with GetDisplayAvatarUrl(), which returns the custom avatar if set, and the default avatar if not set.