fix: pagination not working in DM
Previously, using a command that was returning a paginated response would also return an error "Object reference not set to an instance of an object.".

This commit fixes that issue, which allows pagination to be used on DM interactions.

Also added thumbnail restoration, and cleaned up the syntax a little bit.